fix(snackbar): providerからsnackbarを出力
