+++ packages/wiz-ui-next/src/hooks/use-snackbar.ts
@@ -1,30 +1,6 @@
-import { h, render, ref, VNode } from "vue";
-
-import WizSnackbarController from "@/components/base/snackbar/snackbar-controller.vue";
-import { SnackbarOption } from "@/components/base/snackbar/types";
-
-const options = ref<SnackbarOption[]>([]);
-let snackbar: VNode | null = null;
+import { globalInject, globalKey } from "@/hooks/use-global-provider";
 
 export const useSnackbar = () => {
-  const show = (message: string) => {
-    const created = new Date().toISOString();
-    const deleteSnackbar = () => {
-      options.value = options.value.filter(
-        (options) => options.created !== created
-      );
-    };
-    options.value.push({
-      message,
-      created,
-      delete: deleteSnackbar,
-    });
-  };
-
-  if (!snackbar) {
-    snackbar = h(WizSnackbarController, { options });
-    render(snackbar, document.body);
-  }
-
-  return show;
+  const { snack } = globalInject(globalKey);
+  return snack;
 };
