+++ packages/wiz-ui/src/hooks/use-snackbar.ts
@@ -1,34 +1,6 @@
-import Vue, { ref } from "vue";
-
-import WizSnackbarController from "@/components/base/snackbar/snackbar-controller.vue";
-import { SnackbarOption } from "@/components/base/snackbar/types";
-
-const options = ref<SnackbarOption[]>([]);
-let snackbar: any = null;
+import { globalInject, globalKey } from "@/hooks/use-global-provider";
 
 export const useSnackbar = () => {
-  const show = (message: string) => {
-    const created = new Date().toISOString();
-    const deleteSnackbar = () => {
-      options.value = options.value.filter(
-        (options) => options.created !== created
-      );
-    };
-    options.value.push({
-      message,
-      created,
-      delete: deleteSnackbar,
-    });
-  };
-
-  if (!snackbar) {
-    snackbar = new Vue({
-      ...WizSnackbarController,
-      propsData: { options },
-    });
-    snackbar.$mount();
-    document.body.appendChild(snackbar.$el);
-  }
-
-  return show;
+  const { snack } = globalInject(globalKey);
+  return snack;
 };
